PythonApp/requirements.txt
fastapi
uvicorn
celery[redis]
psycopg2-binary
python-dotenv
SQLAlchemy
pypdf
openai
requests
python-multipart
passlib[bcrypt]
python-jose[cryptography]
alembic